Add support for EnvVar and EnvFromSource in spec
This allows the user to provide environment variables (directly or from
secrets or configmaps) from the custom resource YAML file directly.

These env variables are passed to the containers by the operator when
the resource is installed.

Add the tasksrs example which illustrate how to configure a Java
application that uses a PostgreSQL database.