[#30] Update WildFlyServer Status
Annotate WildFlyServer type with kubebuilder:subresource:status
to ensure that its status is properly registered as a subresource.

If the route kind is not supported, set `DisableHTTPRoute` to false in
the Spec to avoid trying to create the route everytime a requeue
happens.

This fixes https://github.com/wildfly/wildfly-operator/issues/30