[#46] Readiness/liveness scripts
Modify the pod container templates to setup liveness/readiness scripts.

If the SERVER_LIVENESS_SCRIPT env var is defined, the operator will use
it to configure an Exec livenessProbe instead of the default HTTP one.

If the SERVER_READINESS_SCRIPT env var is defined, the operator will use
it to configure an Exec readinessProbe. If it not defines, there is no
readiness probe at all.

This fixes #46.