[#25] Integrate with Prometheus metrics
When ServiceMonitor resource is installed on the cluster,
create a ServiceMonitor to expose WildFly metrics.
The metrics are available on the management port of WildFly (9990)
using the standard Prometheus endpoints (/metrics).
A -admin service is created to expose the management port of WildFly and
is referenced by the ServiceMonitor resource.

This fixes #25.

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>